backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
akshare>=1.12.0
requests>=2.31.0
python-dateutil>=2.8.2
pandas>=2.0.0
numpy>=1.24.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
snownlp>=0.12.3
django-apscheduler>=0.6.2
